refactor(ssa NewCall method)
